requirements.txt
pyserial
opencv-python
requests
imutils
paramiko
paho-mqtt
matplotlib
dash